Migrate types in solend to index.ts
